+++ tortoise/__init__.py
@@ -722,4 +722,4 @@ def run_async(coro: Coroutine) -> None:
         loop.run_until_complete(Tortoise.close_connections())
 
 
-__version__ = "0.15.2"
+__version__ = "0.15.3"